+++ src/routes/auth.js
@@ -5,5 +5,6 @@ const router = express.Router();
 
 router.post('/auth/login', createAuth);
 router.post('/auth/refresh', reissueAccessToken);
+router.get('/auth/callback/:type');
 
 export default router;
